Implement live updating of a cached follower list
AppPreferences now aggregates a follower list and will automatically update whenever there are changes in the follower list of the user. This makes it possible to decouple the follower list document get() operation from the follower list refresh

FollowingListener is a new interface and it handles the callback to the list update mentioned above.

Mood now requires an owner field to be set and is now enforced by the security rules