Re-implement state and situation spinners to have custom adapters.
This commit fixes #6 and fixes #15.

Both EmotionalState and SocialSituation now have custom adapters for use with spinners. Both adapters are "hardcoded" to populate the spinner with the enumeration values. This should be safe, as the spinner is directly defined by the enumerations and should not be reused elsewhere.

DefineMoodDialogFragment now uses the custom adapter in conjunction with the spinner in order to safely and consistently get the item from the spinner. This avoid the coupling with the situation code in the SocialSituation enum. In other words with the old system if any of the situation code or the display name changed, this would break.

Remove SocialSituation OPTIONAL category as to not couple the optional argument with the enumeration definition.