Change the way requests are submitted on the client
This commit combined with security changes fixes #4.

Instead of relying on firestore for creating a unique ID, we set the ID to be the "from" + "to" fields. This way every request is IDd uniquely and can be rejected from the firestore security rules.

In addition, a security rule change has been made that disallows any new requests if the current user already follows the user that is in the request. The error string has been updated to cover both of these cases.